CRM - index Controller

+++ src/main/java/pl/coderslab/Priority/PriorityController.java
@@ -11,12 +11,17 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.SessionAttribute;
+import org.springframework.web.bind.annotation.SessionAttributes;
 
+import pl.coderslab.Commons.Utils.UserRole;
 import pl.coderslab.Priority.Service.PriorityService;
 import pl.coderslab.Priority.dto.PriorityDto;
+import pl.coderslab.User.dto.UserDto;
 
 @Controller
 @RequestMapping("/priority")
+@SessionAttributes("loggedUser")
 public class PriorityController {
 
 	private final PriorityService priorityService;
@@ -27,7 +32,12 @@ public class PriorityController {
 	}
 
 	@GetMapping("/all")
-	public String getAllPriorities(Model model) {
+	public String getAllPriorities(Model model,
+			@SessionAttribute("loggedUser") UserDto loggedUser) {
+		if (loggedUser.getId() == null
+				|| !loggedUser.getUserRole().equals(UserRole.ADMIN)) {
+			return "redirect:/";
+		}
 
 		model.addAttribute("priorities", priorityService.getAll());
 
@@ -35,7 +45,12 @@ public class PriorityController {
 	}
 
 	@GetMapping("/add")
-	public String addNewPriority(Model model) {
+	public String addNewPriority(Model model,
+			@SessionAttribute("loggedUser") UserDto loggedUser) {
+		if (loggedUser.getId() == null
+				|| !loggedUser.getUserRole().equals(UserRole.ADMIN)) {
+			return "redirect:/";
+		}
 
 		model.addAttribute("priority", new PriorityDto());
 
@@ -45,7 +60,12 @@ public class PriorityController {
 	@PostMapping("/add")
 	public String processNewPriority(
 			@Valid @ModelAttribute("priority") PriorityDto dto,
-			BindingResult result) {
+			BindingResult result,
+			@SessionAttribute("loggedUser") UserDto loggedUser) {
+		if (loggedUser.getId() == null
+				|| !loggedUser.getUserRole().equals(UserRole.ADMIN)) {
+			return "redirect:/";
+		}
 		if (result.hasErrors()) {
 			return "priority/priorityForm";
 		}
@@ -56,7 +76,12 @@ public class PriorityController {
 	}
 
 	@GetMapping("/edit/{id}")
-	public String editPriority(@PathVariable("id") Long id, Model model) {
+	public String editPriority(@PathVariable("id") Long id, Model model,
+			@SessionAttribute("loggedUser") UserDto loggedUser) {
+		if (loggedUser.getId() == null
+				|| !loggedUser.getUserRole().equals(UserRole.ADMIN)) {
+			return "redirect:/";
+		}
 
 		model.addAttribute("priority", priorityService.findById(id));
 
@@ -66,7 +91,12 @@ public class PriorityController {
 	@PostMapping("/edit/**")
 	public String processEditPriority(
 			@Valid @ModelAttribute("priority") PriorityDto dto,
-			BindingResult result) {
+			BindingResult result,
+			@SessionAttribute("loggedUser") UserDto loggedUser) {
+		if (loggedUser.getId() == null
+				|| !loggedUser.getUserRole().equals(UserRole.ADMIN)) {
+			return "redirect:/";
+		}
 		if (result.hasErrors()) {
 			return "priority/priorityForm";
 		}
@@ -74,12 +104,17 @@ public class PriorityController {
 
 		return "redirect:/priority/all";
 	}
-	
+
 	@GetMapping("/delete/{id}")
-	public String deletePriority(@PathVariable("id")Long id) {
-		
+	public String deletePriority(@PathVariable("id") Long id,
+			@SessionAttribute("loggedUser") UserDto loggedUser) {
+		if (loggedUser.getId() == null
+				|| !loggedUser.getUserRole().equals(UserRole.ADMIN)) {
+			return "redirect:/";
+		}
+
 		priorityService.deleteFromDb(id);
-		
+
 		return "redirect:/priority/all";
 	}
 }
